+++ include/esp32gfx/raster.hpp
@@ -8,16 +8,16 @@ namespace esp32gfx {
 template<Pixel P>
 void fill_triangle_gouraud(
     Surface<P>& surf,
-    int16_t x0, int16_t y0, Color c0, int16_t z0,
-    int16_t x1, int16_t y1, Color c1, int16_t z1,
-    int16_t x2, int16_t y2, Color c2, int16_t z2);
+    int32_t x0, int32_t y0, Color c0, int16_t z0,
+    int32_t x1, int32_t y1, Color c1, int16_t z1,
+    int32_t x2, int32_t y2, Color c2, int16_t z2);
 
 template<Pixel P>
 inline void fill_triangle_flat(
     Surface<P>& surf,
-    int16_t x0, int16_t y0,
-    int16_t x1, int16_t y1,
-    int16_t x2, int16_t y2,
+    int32_t x0, int32_t y0,
+    int32_t x1, int32_t y1,
+    int32_t x2, int32_t y2,
     Color color,
     int16_t z0 = 0, int16_t z1 = 0, int16_t z2 = 0)
 {
@@ -27,13 +27,13 @@ inline void fill_triangle_flat(
 template<Pixel P>
 void draw_line(
     Surface<P>& surf,
-    int16_t x0, int16_t y0,
-    int16_t x1, int16_t y1,
+    int32_t x0, int32_t y0,
+    int32_t x1, int32_t y1,
     Color color);
 
-extern template void fill_triangle_gouraud(SurfaceRGBA32&, int16_t, int16_t, Color, int16_t, int16_t, int16_t, Color, int16_t, int16_t, int16_t, Color, int16_t);
-extern template void fill_triangle_gouraud(SurfaceGray8&, int16_t, int16_t, Color, int16_t, int16_t, int16_t, Color, int16_t, int16_t, int16_t, Color, int16_t);
-extern template void draw_line(SurfaceRGBA32&, int16_t, int16_t, int16_t, int16_t, Color);
-extern template void draw_line(SurfaceGray8&, int16_t, int16_t, int16_t, int16_t, Color);
+extern template void fill_triangle_gouraud(SurfaceRGBA32&, int32_t, int32_t, Color, int16_t, int32_t, int32_t, Color, int16_t, int32_t, int32_t, Color, int16_t);
+extern template void fill_triangle_gouraud(SurfaceGray8&, int32_t, int32_t, Color, int16_t, int32_t, int32_t, Color, int16_t, int32_t, int32_t, Color, int16_t);
+extern template void draw_line(SurfaceRGBA32&, int32_t, int32_t, int32_t, int32_t, Color);
+extern template void draw_line(SurfaceGray8&, int32_t, int32_t, int32_t, int32_t, Color);
 
 } // namespace esp32gfx
